moved CharacterDisplay dir into components dir
