--- src/CharacterDisplay/CharacterDisplay.css
@@ -1,67 +1,0 @@
-.border {
-  border: 1px solid black;
-}
-
-.main-container {
-  margin: 0 100px;
-  padding: 10px;
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  grid-template-rows: repeat(8, auto);
-  gap: 10px;
-}
-
-.name {
-  grid-column: 1 / 3;
-  grid-row: 1 / 5;
-}
-
-.name h3 {
-  margin-bottom: 0;
-}
-
-.char-info {
-  grid-column: 3 / 7;
-  grid-row: 1 / 5;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: 1fr 1fr;
-}
-
-.char-info-class {
-  grid-column: 1 / 2;
-  grid-row: 1 / 2;
-}
-
-.char-info-level {
-  grid-column: 2 / 3;
-  grid-row: 1 / 2;
-}
-
-.char-info-race {
-  grid-column: 1 / 2;
-  grid-row: 2 / 3;
-}
-
-.ability-score {
-  grid-column: 1 / 2;
-  grid-row: 5 / 9;
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: repeat(6, 1fr);
-  gap: 10px;
-}
-
-.ability {
-  margin: 5px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.modifier {
-    border: 1px solid black;
-    border-radius: 50%;
-    width: 25%;
-}
